Show more event types in GitHubEvents component
Refactor getRecentEvents.ts to use an array of EventProcessors to loop through events, instead of lots of if's.
Store a promise for GitHub events as state in GitHubEvents and then pass it to GitHubEventsList, to avoid having to cache the promise in getRecentEvents().
Include a unique id on each event.
Add word-break: break-word so the description can wrap even with long URLs.
Tweak card styling.
Replace MarkdownContent type, which is no longer in Astro v5.