[Fix #26] Update deploy-pr-build.yml to deploy PR builds to Cloudflare
Get the custom domain from the Pages settings and use it as the site param in the Astro build.
Add a project name env var for the Cloudflare build.
Add a comment with the alias URL for the PR build, and update that comment when new commits are pushed instead of adding new comments each time.